+++ src/main/java/org/hackreduce/models/PrimeRecord.java
@@ -4,7 +4,7 @@ import org.apache.hadoop.io.Text;
 
 public class PrimeRecord {
 
-	int num;
+	long num;
 
 	public PrimeRecord(Text text) throws IllegalArgumentException {
 		this(text.toString());
@@ -13,18 +13,18 @@ public class PrimeRecord {
 	public PrimeRecord(String inputString) throws IllegalArgumentException {
 		String[] attributes = inputString.split(",");
 		try {
-			setNum(Integer.parseInt(attributes[0]));
+			setNum(Long.parseLong(attributes[0]));
 		} catch (NumberFormatException e) {
 			throw new IllegalArgumentException("Input string contained an unknown number value that couldn't be parsed", e);
 		}
 		
 	}
 
-	public int getNum() {
+	public long getNum() {
 		return num;
 	}
 
-	public void setNum(int value) {
+	public void setNum(long value) {
 		this.num = value;
 	}
 }
